moved menu.dart to menu folder
